+++ src/main/java/orm/dsl/dialect/h2/H2CreateTableImpl.java
@@ -1,4 +1,4 @@
-package orm.dsl.ddl.dialect.h2;
+package orm.dsl.dialect.h2;
 
 import orm.TableEntity;
 import orm.TableField;
