+++ src/main/java/orm/dsl/step/ddl/DDLTableIfNotExistStep.java
@@ -1,4 +1,4 @@
-package orm.dsl.ddl;
+package orm.dsl.step.ddl;
 
 import orm.dsl.QueryBuilder;
