+++ src/main/java/orm/dsl/step/ddl/DropTableStep.java
@@ -1,4 +1,4 @@
-package orm.dsl.ddl;
+package orm.dsl.step.ddl;
 
 public interface DropTableStep extends DDLTableIfNotExistStep {
